requirements.txt
# === OpenKB (Wiki 시스템) ===
openkb>=0.1.0

# === LLM 연동 ===
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
httpx>=0.27.0

# === Agent 및 유틸리티 ===
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# === 웹 UI ===
streamlit>=1.35.0

# === 문서 변환 ===
python-docx>=1.0.0          # DOCX → TXT
pymupdf>=1.24.0             # PDF → TXT
